manualCmr: time out trial after numSweeps done
git-svn-id: svn://132.239.158.177/projects/ratrix/trunk@2508 d291aab3-510b-4049-a6d1-fe65b8d52d70

+++ classes/protocols/stimManagers/@manualCmrMotionEyeCal/calcStim.m
@@ -45,8 +45,8 @@ discrimStim.startFrame=0;
 discrimStim.autoTrigger=[];
 onlyOnce=find(ismember(stimulus.stateValues,{'initialize','done'}));
 discrimStim.framesUntilTimeout=hz*...
-    (sum(stimulus.stateDurationValues(setdiff(1:length(stimulus.stateDurationValues),onlyOnce)))...
-    *stimulus.numSweeps+sum(stimulus.stateDurationValues(onlyOnce)));
+    (sum(double(stimulus.stateDurationValues(setdiff(1:length(stimulus.stateDurationValues),onlyOnce))))...
+    *double(stimulus.numSweeps)+sum(double(stimulus.stateDurationValues(onlyOnce))));
 preRequestStim=[];
 preRequestStim.stimulus=interTrialLuminance;
 preRequestStim.stimType='loop';
